Creating custom linux bootentries (#1221)
* util/bootentries-fix.robot: Add PoC for adding bootentries


* options-lib_dcu.robot: Fix faulty comparison of bootid to empty


* lib/flash.robot: Add bootorder replacement when flashing fw and tesing via ssh


* Execute Reboot Command: Use BOOTED_OS_ID to determine reboot type


* novacustom-v540tu.robot: Resture Ubuntu as default boot os


* lib/custom_bootentries.robot: Add library for managing bootorder


* lib/flash.robot: Migrate smmstore when flashing without access to serial


* util/bootentries-fix.robot: Clean up the suite


* util/bootentries-fix.robot: Move to self-tests/bootentries-persistence.robot


* util/basic-platform-setup: Add BPS005.002 for creating defaultboot entires


* BPS009: Add suite that adds DEFAULT_BOOT_OS_ID entry


* bootentries-persistence.robot: Make it depend on basic platform setup done


* lib/custom_bootentries.robot: Make sure esp are skipped


* environment-test-ids.py: Extend Windows bootmenu name


* BPS009: make sure windows bootentry is created


* Execute Shutdown Command: Skip if no POWER_CTRL


---------

+++ keywords.robot
@@ -799,11 +799,19 @@ Restore Initial DUT Connection Method
     END
 
 Execute Shutdown Command
-    [Documentation]    Windows shutdown keyword, checks power LED state where available.
+    [Documentation]    Shutdown keyword, checks power LED state where available.
     ...    Depends on existing SSH connection to DUT, restores initial connection method
     ...    after power loss.
-    VAR    ${DUT_CONNECTION_METHOD}=    SSH    scope=GLOBAL
-    Execute Command In Terminal    shutdown /s /f /t 0
+    IF    '${POWER_CTRL}' == 'none'
+        # Skip if there is no way to boot the device back up after shutdown.
+        RETURN
+    END
+    IF    '${BOOTED_OS_ID}'.startswith("2")
+        Execute Command In Terminal    shutdown 0
+    ELSE IF    '${BOOTED_OS_ID}' == '${ENV_ID_WINDOWS}'
+        VAR    ${DUT_CONNECTION_METHOD}=    SSH    scope=GLOBAL
+        Execute Command In Terminal    shutdown /s /f /t 0
+    END
     IF    '${CHECK_POWER_LED_SUPPORT}' == '${TRUE}'
         ${loop_iterations}=    Evaluate    ${WINDOWS_SHUTDOWN_AWAITING_SECONDS} * 2
         FOR    ${i}    IN RANGE    ${loop_iterations}
@@ -821,8 +829,8 @@ Execute Poweroff Command
 
 Execute Reboot Command
     [Documentation]    Executes reboot command in given os
-    [Arguments]    ${os}=linux    ${assume_correct_boot}=${False}
-    IF    '${os}' == 'linux'
+    [Arguments]    ${os}=${BOOTED_OS_ID}    ${assume_correct_boot}=${False}
+    IF    '${os}' == 'linux' or "${os}".startswith('2')
         # if the OS cannot be chosen from the bootmanager and rebooting
         # always boots the default one
         IF    '${OPTIONS_LIB}' == 'options-lib_dcu' and ${assume_correct_boot} == ${False}
@@ -831,7 +839,7 @@ Execute Reboot Command
             VAR    ${BOOTED_OS_ID}=    ${BOOTED_OS_ID}    scope=GLOBAL
         END
         Write Into Terminal    reboot
-    ELSE IF    '${os}' == 'windows'
+    ELSE IF    '${os}' == 'windows' or '${os}'.startswith('3')
         Write Into Terminal    shutdown /r /f /t 0
     ELSE
         Fail    Unknown OS: ${os} given as an argument.
